+++ constants.go
@@ -289,6 +289,7 @@ const (
 	NetworkTCP  = "tcp"
 	NetworkTCP4 = "tcp4"
 	NetworkTCP6 = "tcp6"
+	NetworkUnix = "unix"
 )
 
 // Compression types